check windowsize at app launch
