+++ src/components/main/main.tsx
@@ -2,8 +2,11 @@ import React, { FunctionComponent, useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 
 import { initMobroClient } from "../../store/moBro/mobroSlice";
-import { initSelector } from "../../store/moBro/mobroSelectors";
-import { useWindowSize } from "../../hooks";
+import {
+  initSelector,
+  windowSizeSelector,
+} from "../../store/moBro/mobroSelectors";
+
 import { CpuTemperature, CpuUsage, CpuDetails } from "../cpu";
 import { GpuDetails, GpuUsage, GpuTemperature, GpuVram } from "../gpu";
 import { RamUsage } from "../ram/ram";
@@ -15,8 +18,8 @@ import { StyledMainContainer } from "./main.style";
 
 export const Main: FunctionComponent = () => {
   const dispatch = useDispatch();
-  const windowSize = useWindowSize();
   const init = useSelector(initSelector);
+  const windowSize = useSelector(windowSizeSelector);
 
   useEffect(() => {
     dispatch(initMobroClient());
